+++ ComputingTheConvexHullOnGpu/Point.cs
@@ -2,10 +2,10 @@ namespace ComputingTheConvexHullOnGpu
 {
     public readonly struct Point
     {
-        public double X { get; }
-        public double Y { get; }
+        public float X { get; }
+        public float Y { get; }
 
-        public Point(double x, double y)
+        public Point(float x, float y)
         {
             X = x;
             Y = y;
